+++ ImTK/UI/Element/Drawer/Drawers/StringDropdownDrawer.cs
@@ -1,0 +1,9 @@
+namespace ImTK.UI
+{
+    public class StringDropdownDrawer : DropdownDrawer<string>
+    {
+        public StringDropdownDrawer()
+        {
+        }
+    }
+}